rename .*.env.ts -> env.ts
