+++ test/pending-sale/pending-sale.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect } from 'vitest';
 import { User } from '../user';
 import { whitelistTier0, whitelistTier1 } from '../well-known-users';
-import env from './.env.pending-sale';
+import env from './env';
 
 describe('pending sale', () => {
   test('check salesSettings', async () => {
