+++ test/public-sale/env.ts
@@ -1,7 +1,6 @@
-import defaultEnv from '../.env.default';
+import defaultEnv from '../default-env';
 
 export default {
   ...defaultEnv(),
-  sale: 'variant { supply = 5 }',
   publicSaleStart: BigInt(Date.now()) * 1_000_000n,
 };
