+++ test/sold-out/env.ts
@@ -1,6 +1,7 @@
-import defaultEnv from '../.env.default';
+import defaultEnv from '../default-env';
 
 export default {
   ...defaultEnv(),
+  sale: 'variant { supply = 5 }',
   publicSaleStart: BigInt(Date.now()) * 1_000_000n,
 };
